refactor: move sync package to internal
