+++ internal/sync/merge.go
@@ -1,4 +1,4 @@
-package syncmd
+package sync
 
 import (
 	"bytes"
