+++ internal/sync/sync.go
@@ -1,4 +1,4 @@
-package syncmd
+package sync
 
 import (
 	"errors"
